+++ api/agent/context_usage.py
@@ -10,7 +10,7 @@ import tiktoken
 from PIL import Image
 
 from agent.prompts import get_system_prompt_parts
-from api.session_manager import SessionState
+from api.session.manager import SessionState
 
 _ENCODING = None
