+++ tests/test_memory/test_user_init.py
@@ -1,6 +1,6 @@
 """memory/user_init.py 测试。"""
 
-import memory.user_init as user_init
+import api.memory.user_init as user_init
 
 
 class TestEnsureUserMd:
