+++ tools/memory/tool_merge_memories.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from pydantic import BaseModel, Field
 
 from tools.base import ToolBase, format_error, format_success
-from memory.memory_manager import MAX_DESC_LENGTH
+from api.memory.manager import MAX_DESC_LENGTH
 
 
 class MergeMemoriesInput(BaseModel):
@@ -66,7 +66,7 @@ class MergeMemoriesTool(ToolBase):
         if not reason:
             return format_error("reason 不能为空，请说明合并原因")
 
-        from memory.memory_manager import MemoryManager
+        from api.memory.manager import MemoryManager
 
         mm = MemoryManager(yaml_file=str(MEMORY_PATH))
         try:
