+++ tools/memory/tool_read_memories.py
@@ -41,7 +41,7 @@ class ReadMemoriesTool(ToolBase):
         if not MEMORY_PATH.exists():
             return format_error("记忆文件不存在")
 
-        from memory.memory_manager import MemoryManager
+        from api.memory.manager import MemoryManager
 
         mm = MemoryManager(yaml_file=str(MEMORY_PATH))
         items = mm.show()
