+++ tools/sub_agent/tool_call_sub_agent.py
@@ -7,7 +7,7 @@ import traceback
 
 from pydantic import BaseModel, Field
 
-from api import interaction
+from api.agent import interaction
 from tools.base import ToolBase, format_success, format_error
