+++ tools/system/tool_python.py
@@ -6,7 +6,7 @@ import sys
 
 from pydantic import BaseModel, Field
 
-from api import interaction
+from api.agent import interaction
 from tools.base import ToolBase, format_error, format_success, get_safe_builtins
 
 # 模块级常量：安全 builtins 只构造一次
